use size_t for vector index

+++ src/BKtree_utils.hpp
@@ -14,16 +14,16 @@ static int levenshtein_distance(std::string str1, std::string str2){
   size_t m = str1.length(), n = str2.length();
   
   int** dn = new int*[m + 1];
-  for(int i = 0; i < m + 1; ++i) {
+  for(size_t i = 0; i < m + 1; ++i) {
     dn[i] = new int[n + 1];
   }
   
-  for (int i = 0; i < m + 1; ++i) {
-    for (int j = 0; j < n + 1; ++j) {
+  for (size_t i = 0; i < m + 1; ++i) {
+    for (size_t j = 0; j < n + 1; ++j) {
       if (i == 0) {
         dn[0][j] = j; // deletions at the start of str2
         
-      } else if (j==0) {
+      } else if (j == 0) {
         dn[i][0] = i; // deletions at the start of str1
         
       } else if (str1[i-1] == str2[j-1]) { // match of str1[i-1] and str2[j-1]
@@ -31,16 +31,16 @@ static int levenshtein_distance(std::string str1, std::string str2){
         
       } else { // mismatch between str1[i-1] and str2[j-1] -> find minimal source
         dn[i][j] = 1 + std::min(
-          dn[i-1][j], // deletion in str1
-                 std::min(dn[i][j-1], // deletion in str2
-                          dn[i-1][j-1]) // mismatch
+          dn[i-1][j],          // deletion in str1
+          std::min(dn[i][j-1], // deletion in str2
+          dn[i-1][j-1])        // mismatch
         );
       }
     }
   }
   
   int d = dn[m][n];
-  for (int i = 0; i < m + 1; ++i) {
+  for (size_t i = 0; i < m + 1; ++i) {
     delete [] dn[i];
   }
   delete [] dn;
